rename feature python dir
