+++ python/runtime/feature/__init__.py
@@ -11,4 +11,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from runtime.feature_derivation.feature_derivation import infer_feature_columns
+from runtime.feature.derivation import infer_feature_columns
